+++ src/main/java/org/example/psychologicalcounseling/repository/AccountRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface AccountRepository extends JpaRepository<Account, Integer> {
+public interface AccountRepository extends JpaRepository<Account, Long> {
     // 根据 aid 查询 password
     @Query("SELECT a.password FROM Account a WHERE a.aid = :aid")
-    String findPasswordByAid(@Param("aid") int aid);
+    String findPasswordByAid(@Param("aid") Long aid);
 }
